requirements.txt
torch>=1.9.0
numpy>=1.19.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
pyyaml>=5.4.0
tqdm>=4.60.0
scipy>=1.6.0
pandas>=1.3.0
transformers>=4.0.0

# Optional: For full TRAK implementation
# traker>=0.1.0